Fix multiple loading of custom shell files
Previously, a .zsh or .bash file could be loaded multiple times by the
startup script. This was due to how the $exts local variable was updated
within the loop over plugins.

This was fixed by setting the $exts variable outside of the loop, so
that it wouldn't acquire multiple extensions, resulting in the file
being loaded multiple times.