Consolidate irb and rvm plugins into ruby plugin
There was an issue where the irbrc specified by the irb plugin was not
being used because sourcing the rvm scripts caused the rvm-specified
version irbrc to be used. Since I have no intention of using a
version-specific irbrc, I need to merge the two together so that I can
guarantee the order of execution (i.e. source the rvm script and /then/
set IRBRC environment variable to point to custom configuration).