Don't use tmp directory when installing Dot
Previously, we were creating a file in the /tmp directory before moving
it to the appropriate location in the home directory (so we could do a
diff beforehand). The problem is that CentOS doesn't seem to have
$TMPDIR defined, so it thinks we're trying to write to the root
directory. A simple fix is to change the temporary name of the file to
something unlikely to conflict and just create it in $PWD.