Don't generate tag entries for Python imports
Previously, jumping to a tag definition in Python files was usually a
pain since it would always ask you which entry you wanted to jump to (if
the function appeared in multiple `import` statements, for example).

Since these entries are seemingly useless, adding a line to the ctags
file to ignore them makes jumping to definition much easier.