Make `git ctags` ignore warnings
Previously if `ctags` output any warnings you would see this when
running `git ctags`. This was rather annoying, as the warnings weren't
terribly useful.

This commit changes the ctags script to capture all STDERR to a file,
and only display that file if the process executed with a non-zero
status (the theory being that it's only interesting if it causes a hard
failure).