Display action item in `git fresh` when rebase fails
Previously, if a conflict occurred during a rebase in `git fresh`, there
was no hint as to what to do to solve the problem. To make problem
solving easier, a simple message was added to the error to tell the user
the exact `git rebase` command to use fix the issue themselves.