Coalesce {lib,env}.sh files into plugin.sh
In the original design of Dot, it was believed that lib and env files
would grow to large sizes, and so it made sense to split them into two
files to separate concerns.

After having used Dot for a while now, this does not seem to be the
case. Therefore, to simplify things and speed up load time slightly,
this commit condenses the lib/env files into a single plugin.* file.