Don't initialize colours when installing Dot
Previously, Dot's bootstrap/init.sh file would call init_colours
any time it was included.

This resulted in a problem where running ./install would display an
error about an invalid option for `typeset`. This was due to the fact
that the install script was being run as `sh` but the environment
variables made the script think it was running as `zsh`. However, since
the shell was actually `sh`, the wrong version of `typeset` was being
called, resulting in the error message being shown.

The easiest solution was to only initialize the colours in the bootstrap
`startup.sh` script.