Add `git track` alias
Given the automatic branch tracking that I've set up, there are times
when I've built branches off of other branches which eventually get
merged into master, and so it doesn't make sense to track them anymore.

The `track` alias allows for the easy reassignment of a branch's
tracking branch. When on the branch you want to reassign, just run
`git track <tracking-branch>`.