Add `git tracking` alias
Add `tracking` to list of git aliases so that the remote tracking branch
of a branch is easily determined. This adds a script which takes a
branch name as a parameter and returns the branch (remote or otherwise)
it is tracking. If no branch name is provided, it assumes the current
branch was intended.