Add support for per-machine `.gitconfig.local` file
In order to separate my work email from my personal email when
developing on their respective projects, it would be convenient if I
could set the default email to use on a per-machine basis without having
to check it in to source control.

Using the [include] directive in the `.gitconfig` file allows for this, so
now all per-machine configurations can go in `~/.gitconfig.local`.

This also has the added benefit of me not having to check in my email to
source control. Even if I forget to create a .gitconfig.local file,
`git` will automatically remind me the first time I try to commit.