Add newline in MySQL prompt to fix prompt width
MySQL counts the characters used to do colour coding in the prompt
width, preventing the command a user enters from taking up the full
width of the terminal since MySQL thinks the prompt is wider than it
actually is.

Adding a newline avoids this entirely by resetting the line that the
command appears on to not have any colours on it.