Add DSL for Homebrew formulae and casks
In order to extend Dot to support installing and managing GUI applications,
add a DSL `formula` and `cask` to specify that a formula or cask should
be installed with Homebrew.

This may be Mac-specific, but it is hopefully implemented in such a way
that similar DSLs could be implemented for other OS vendors (should I
ever decide I want to support both).