Enable pbcopy/pbpaste in tmux
Since `tmux` starts its server using `daemon`, it loses connection to
the user's namespace, which makes it unable to access the user's
clipboard.

This is solvable by using a tool called `reattach-to-user-namespace`,
with is installable via Homebrew. By changing the `default-command` of
`tmux` to check for the existence of the program and starting up a shell
using it, we can reattach the shell to the user namespace.

As a fallback, if `reattach-to-user-namespace` not installed we just
open the user's shell based on `$SHELL`.