Improve tmux helper script
Adds a check so that we don't attempt to assign the SSH_AUTH_SOCK
variable in the case it's not set to anything. This prevents a situation
from occurring where someone else logging into your system (e.g. someone
you want to pair program with) who does not have your private SSH key
will clobber your `SSH_AUTH_SOCK` variable when they attach to your
client by running the `tmux` helper script.

Also changed one of the commands to use the `-A` flag to avoid a more
complicated command with an `||`.