Change install script to work without git
Previously, Dot's install script assumed `git` already existed on the
system.

This presents a problem on "blank slate" new Macs, since they don't come
with `git` out of the box.

The solution is to `curl` Homebrew as a tarball and extract it to
`/usr/local`, then install `git` via Homebrew.

We don't need a GCC compiler for this to work, so the `install_package`
helper and the associated line that invoked it to install `gcc` was
removed.

The README was updated to explain how to install on a blank machine.