Add Clipper plugin
Installs and loads `clipper` (https://github.com/wincent/clipper) which
makes working with the clipboard on remote systems easier.

In the process, adds support to the `formula` DSL to load/unload a
particular plist file as part of installing/uninstalling the formula.

The end result is that `clipper` is downloaded, built from source,
and loaded running as a daemon when it is installed.