Symlink hushlogin
Previously we were creating an empty file `.hushlogin` in the user's
home directory to disable the welcome message on login.

In order to make it easier to tell which files are managed by Dot,
switch it to use a symlink, since symlinks appear in a different colour
when you list the home directory using `ls`.