Hard reset master branch on repo initialization
If Dot was installed via tarball, there won't be any `.git` directory in
the repository, so we need to create one.

However, one step that was missed was to also set the `master` branch
via a soft reset (soft so that in the event the user has local changes
we don't overwrite them).