Ensure git repository when installing/updating
When installing/updating Dot, we want to make sure the git repository
has been initialized. For installation, we wait until after the
installation so that Dot gets a chance to install `git` in case it
doesn't exist. For updating, we expect `git` to already be installed and
so ensure it is a git repository before updating (which is really where
it's important to do so).