Don't try to detect if git is installed
On a new Mac with Mavericks installed, there is a binary at /usr/bin/git
that appears to be git, but is really a stub telling you to install
XCode.

Since this makes it slightly more annoying to determine if git is
installed, don't bother detecting if it is installed to see if we can
just clone the repo: install by downloading and extracting the tarball
regardless, since we'll still init the git repository when we're done.